requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.3
django-environ==0.11.2
django-filter==24.1
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.15.0
drf-spectacular==0.27.1
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
